fix conformence

+++ pkg/baloum/map_prog_array_test.go
@@ -72,7 +72,7 @@ func TestTailCall(t *testing.T) {
 
 	var ctx StdContext
 	code, err := vm.RunProgram(&ctx, "test/tail_call")
-	assert.Equal(t, 72, code)
+	assert.Equal(t, int64(72), code)
 	assert.Nil(t, err)
 
 	data, err := vm.Map("data").Lookup(uint64(0))
